feat: Add basic lexical analysis products to the llvm ir grammar
This grammar is the basic lexical grammar for llvm ir minus keywords
and operator symbols which are usually implicitly added as part of the
syntax rules.

For now a few operator symbols have been added here just to make the grammar
work for simple examples.

The rest of the grammar is coming soon.