feat: eleminate left recursion from grammar
There is still an issue with pointers and left recursin.

The grammar needs to be refactored to eliminate lots of predefined
string literals like "br" so that a general string literatl is ues there.
It will cause unecessary lookahead in the scanner otherwise.

Also, there are many productions that should be deleted and put under
a more general umbrella. This version of the grammar was an almost
literal translation of another grammar.